move generic event utils to testutil
